% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changedfnames.R
\name{changedfnames}
\alias{changedfnames}
\title{changedfnames Function}
\usage{
changedfnames(df, namesin, namesout)
}
\arguments{
\item{df}{Name of the input dataframe}

\item{namesin}{List of current names}

\item{namesout}{List of new names. Must be of equal length as namesin.}
}
\description{
This function receives a dataframe, a list of current names, and a list of equal length
of new names.  It returns a dataframe with the new names, or stops with an error message
if it can not do so.
}
\examples{
changedfnames(df=data.frame(a=c(1,2),b=c(3,4)), namesin=c("a","b"), namesout=c("c","d"))
}
\keyword{changedfnames}

