% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_geocode.R
\name{g_geocode}
\alias{g_geocode}
\title{g_geocode Function}
\usage{
g_geocode(addr = NULL, city = NULL, state = NULL, api_key = NULL,
  retrytimes = 5)
}
\arguments{
\item{city}{The city.}

\item{state}{The state.}

\item{api_key}{The Google api_key.  If NULL, a default value from getcredentials will be used.}

\item{retrytimes}{Number of times to try if you do not get a 200 return code}

\item{address}{The street address.}
}
\description{
This function allows you retrieve the latitude and longitude by using a Google API
}
\examples{
g_geocode(addr="222 Malone St", city="Houston", state="TX")
}
\keyword{g_geocode}

