% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcredentials.R
\name{getcredentials}
\alias{getcredentials}
\title{getcredentials Function}
\usage{
getcredentials(system = NA, filepathname = "~/credentials/credentials.csv",
  quiet = FALSE)
}
\arguments{
\item{system}{Name of the desired system for which you'd like to retrieve the credentials. No default.}

\item{filepathname}{Concatenated file and path name to the CSV. Defaults to "~/credentials/credentials.csv"}

\item{quiet}{If TRUE suppresses output from this function. Defaults to FALSE.}
}
\description{
This function allows you to access credentials stored in a csv file of your chosing.
The benefit is that you can store your credentials in a separate location from your code.
The function returns a dataframe if successful, or FALSE if not.
}
\examples{
getcredentials(system="masterdatabase", filepathname="C:/lockbox/passwords.txt")
}
\keyword{getcredentials}

