% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpackage.R
\name{loadpackage}
\alias{loadpackage}
\title{loadpackage Function}
\usage{
loadpackage(pname, quiet = FALSE)
}
\arguments{
\item{pname}{Name of the desired package. No default.}

\item{quiet}{If TRUE suppresses output from this function. Defaults to FALSE.}
}
\description{
This function allows you to load packages cleanly.  They will be downloaded from 
CRAN (if needed) and instantiated into memory. You will know this is successful
if the return value is TRUE.  FALSE indicates failure.
}
\examples{
loadpackage(pname="ggplot2", quiet=FALSE)
}
\keyword{loadpackage}

