% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushtodomo.R
\name{pushtodomo}
\alias{pushtodomo}
\title{pushtodomo Function}
\usage{
pushtodomo(df, datasetname, api_key = NULL, organization = "KonoAnalytics",
  createdataset = FALSE)
}
\arguments{
\item{df}{The data frame you would like to push  to Domo.}

\item{datasetname}{The name of the data set you would like to push to Domo.}

\item{api_key, }{The api key issued by Domo.}

\item{organization}{The organization as determined by Domo.}

\item{createdataset}{TRUE if you want to create a new data set, FALSE otherwise}

\item{dbname}{The name of the database.}
}
\description{
This function allows you to execute a mysql statement against a mysql database. The only required parameter is
the mysql statement.  All others may be NULL.
The function returns a dataframe.
}
\examples{
pushtodomo(df=data.frame(a=c(1,2),b=c(3,4)), datasetname="Test Dataset", createdataset=TRUE)
}
\keyword{pushtodomo}

