% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmysql.R
\name{runmysql}
\alias{runmysql}
\title{runmysql Function}
\usage{
runmysql(statement, con = NULL, user = NULL, password = NULL,
  host = NULL, dbname = NULL)
}
\arguments{
\item{statement}{The MYSQL statement you would like to execute.}

\item{con}{A connection of type MySQLConnection as described in theRMySQL package. If NULL, one will be created.}

\item{user}{The user id of the database.}

\item{password}{The password of the dataase.}

\item{host}{Where the database resides.}

\item{dbname}{The name of the database.}
}
\description{
This function allows you to execute a mysql statement against a mysql database. The only required parameter is
the mysql statement.  All others may be NULL.
The function returns a dataframe.
}
\examples{
runmysql(statement = "select * from tbl_main")
}
\keyword{runmysql}

